Make Cmd work like PV (method forwarding to device)
